04_agent_executor/agent_executor/pyproject.toml
[project]
name = "agent-executor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.9",
    "uvicorn>=0.34.3",
]
